+++ R/crew_client.R
@@ -18,10 +18,10 @@
 #' @param seconds_interval Number of seconds between
 #'   polling intervals waiting for certain internal
 #'   synchronous operations to complete,
-#'   such as checking `mirai::status()`
+#'   such as checking `mirai::info()`
 #' @param seconds_timeout Number of seconds until timing
 #'   out while waiting for certain synchronous operations to complete,
-#'   such as checking `mirai::status()`.
+#'   such as checking `mirai::info()`.
 #' @param serialization Either `NULL` (default) or an object produced by
 #'   [mirai::serial_config()] to control the serialization
 #'   of data sent to workers. This can help with either more efficient
@@ -328,7 +328,7 @@ crew_class_client <- R6::R6Class(
       .started <<- FALSE
       invisible()
     },
-    #' @description Get the counters from `mirai::status()`.
+    #' @description Get the counters from `mirai::info()`.
     #' @return A named integer vector of task counts
     #'   (awaiting, executing, completed) as well as the number of
     #'   worker connections.